second commit'

+++ Day 12/homework12.py
@@ -1,6 +1,6 @@
 #1
-budget = float(input("Enter your budget: "))
-item_cost = float(input("Enter the cost of the desired item: "))
+budget = int(input("Enter your budget: "))
+item_cost = int(input("Enter the cost of the desired item: "))
 if budget >= item_cost:
     print("You can buy the item.")
 else:
